requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
werkzeug==2.3.7
google-generativeai==0.3.1
python-dotenv==1.0.0
flask-wtf==1.2.1
email-validator==2.1.0
flask-migrate==4.0.5
pillow==10.1.0
pydub==0.25.1
python-multipart==0.0.6